added json responses to categories and fixed helper relay
