+++ server/server.js
@@ -2,7 +2,7 @@ import * as dotenv from 'dotenv';
 import express from "express";
 
 import lenses from './src/endpoints/lenses.js';
-import categorylenses from './src/endpoints/categorylenses.js';
+import categorylenses from './src/endpoints/category/lenses.js';
 import top from './src/endpoints/top.js';
 import unlock from './src/endpoints/unlock.js';
 import categories from './src/endpoints/categories.js';
@@ -20,7 +20,7 @@ const serverPort = process.env.PORT;
 const app = express();
 
 app.use(express.json());
-app.use(/vc/v1/explorer/lenses, lenses);
+app.use('/vc/v1/explorer/lenses', lenses);
 app.use('/vc/v1/explorer/category/lenses', categorylenses);
 app.use('/vc/v1/explorer/top', top);
 app.use('/vc/v1/explorer/unlock', unlock);
