+++ server/src/json/character.json
