+++ server/src/json/color_effect.json
