+++ server/src/json/funny.json
